+++ packages/showcases/core/csf/components/overlays/snackbar.csf.js
@@ -25,7 +25,7 @@ export const argTypes = {
   },
   actionLabel: {
     type: { name: 'string', required: false },
-    describe: 'Label of the action. If set, it display action button',
+    describe: 'Label of the action. If set, it displays action button',
     defaultValue: 'Action',
     control: {
       type: 'text',