+++ packages/sources/svelte/src/components/navigation/VtmnBreadcrumb/test/VtmnBreadcrumbWithSlot.svelte
@@ -1,0 +1,10 @@
+<script>
+  import VtmnBreadcrumb from '../VtmnBreadcrumb.svelte';
+  import VtmnBreadcrumbItem from '../../VtmnBreadcrumbItem/VtmnBreadcrumbItem.svelte';
+</script>
+
+<VtmnBreadcrumb {...$$restProps}>
+  <VtmnBreadcrumbItem>Item 1</VtmnBreadcrumbItem>
+  <VtmnBreadcrumbItem>Item 2</VtmnBreadcrumbItem>
+  <VtmnBreadcrumbItem>Item 3</VtmnBreadcrumbItem>
+</VtmnBreadcrumb>